backend/requirements.txt
django-mptt==0.9.1
factory-boy==2.11.1
mock==2.0.0
pyhamcrest==1.9.0
django-cte==1.1.4
djangorestframework
django-cors-headers
django-guardian
django<2.2.0
django-river==3.0.0